implemented rbac and other functions

+++ library-management.py
@@ -8,6 +8,9 @@ def userdb_connection():
 def bookdb_connection():
     return sqlite3.connect("Database/bookdb.db")
 
+def permissiondb_connection():
+    return sqlite3.connect("Database/rbac.db")
+
 def create_database_tables():
     userconn = userdb_connection()
     bookconn = bookdb_connection()
@@ -19,7 +22,8 @@ def create_database_tables():
                             username TEXT NOT NULL,
                             email TEXT NOT NULL,
                             password BLOB NOT NULL,
-                            register_date TEXT DEFAULT CURRENT_TIMESTAMP
+                            register_date TEXT DEFAULT CURRENT_TIMESTAMP,
+                            role_id INTEGER
                             )
                         ''')
     
@@ -65,6 +69,23 @@ def add_book():
 def register():
     userconn = userdb_connection()
     user_cursor = userconn.cursor()
+
+    user_cursor.execute('''CREATE TABLE IF NOT EXISTS users (
+                            id INTEGER PRIMARY KEY AUTOINCREMENT,
+                            username TEXT NOT NULL,
+                            email TEXT NOT NULL,
+                            password BLOB NOT NULL,
+                            register_date TEXT DEFAULT CURRENT_TIMESTAMP,
+                            role_id INTEGER
+                            )
+                        ''')
+    user_cursor.execute('''CREATE TABLE IF NOT EXISTS roles (
+                            id INTEGER PRIMARY KEY,
+                            role_name TEXT UNIQUE
+                            )
+                        ''')
+    user_cursor.execute("INSERT INTO roles (role_name) VALUES ('admin'), ('user')")
+    roles()
 
     print("Willkommen! Du befindest dich im Registrierungsprozess!")
     username = input("Benutzername: ")
@@ -139,11 +160,161 @@ def rent_book():
 
 
 def return_book():
-    pass
+    book_con = bookdb_connection()
+    cursor_book = book_con.cursor()
+
+    author = input("Gib den Autor ein: ")
+    title = input("Gib den Titel ein: ")
+
+    cursor_book.execute("SELECT * FROM books WHERE author = ? AND title = ?", (author, title))
+    result = cursor_book.fetchone()
+    
+    if not result:
+        print("Keine Bücher gefunden!")
+    else:
+        for book in result:
+            print(f"ID: {book[0]}, Titel: {book[1]}, Autor: {book[2]}")
+        try:
+            book_id = int(input("Gib die ID ein: "))
+            cursor_book.execute("UPDATE books SET status = ? WHERE id = ?", ('verfügbar', book_id))
+            book_con.commit()
+        except ValueError:
+            print("Falsche Eingabe!")
+
+    book_con.close()
 
 
 def delete_book():
+    book_con = bookdb_connection()
+    cursor_book = book_con.cursor()
+
+    search = input("Möchtest du nach dem Autor(1) oder Titel(2) suchen: ")
+    if search == '1':
+        author = input("Gib den Namen des/der Autor/in ein: ")
+        cursor_book.execute("SELECT * FROM books WHERE author = ?", ('%' + author + '%',))
+    elif search == '2':
+        title = input("Gib den Titel des Buches ein: ")
+        cursor_book.execute("SELECT * FROM books WHERE title = ?", ('%' + title + '%',))
+    else:
+        print("Ungültige Eingabe!")
+        return
+
+    result = cursor_book.fetchall()
+
+    if not result:
+        print("Keine Bücher gefunden!")
+    else:
+        for book in result:
+            print(f"ID: {book[0]}, Titel: {book[1]}, Autor: {book[2]}, Verfügbarkeit: {book[3]}")
+        try:
+            book_id = int(input("Gib die ID ein: "))
+            cursor_book.execute("DELETE FROM books WHERE id = ?", (book_id,))
+            book_con.commit()
+        except ValueError:
+            print("Falsche Eingabe!")
+    
+    book_con.close()
+
+
+def assign_role(username, role):
+    permissionconn = permissiondb_connection()
+    cursor_per = permissionconn.cursor()
+
+    cursor_per.execute('''CREATE TABLE IF NOT EXISTS user_roles (
+                            username TEXT NOT NULL,
+                            role_name TEXT NOT NULL,
+                            FOREIGN KEY (role_name) REFERENCES roles(role_name)
+                            )
+                       ''')
+    
+    cursor_per.execute("INSERT INTO user_roles (username, role_name) VALUES (?,?)", (username, role))
+    permissionconn.commit()
+    permissionconn.close()
+
+def roles():
+    permissioncon = permissiondb_connection()
+    cursor_per = permissioncon.cursor()
+
+    cursor_per.execute('''CREATE TABLE IF NOT EXISTS permissions(
+                   id INTEGER PRIMARY KEY,
+                   permission_name TEXT UNIQUE
+                   )
+                ''')
+    
+    cursor_per.execute('''CREATE TABLE IF NOT EXISTS roles (
+                   id INTEGER PRIMARY KEY,
+                   role_name TEXT UNIQUE
+                   )
+                ''')
+    
+    cursor_per.execute("INSERT INTO roles (role_name) VALUES ('admin'), ('user')")
+    
+    cursor_per.execute("INSERT INTO permissions (permission_name) VALUES ('create'), ('read'), ('update'), ('delete')")
+
+    cursor_per.execute('''CREATE TABLE IF NOT EXISTS role_permissions (
+                   role_id INTEGER,
+                   permission_id INTEGER,
+                   FOREIGN KEY(role_id) REFERENCES roles(id),
+                   FOREIGN KEY(permission_id) REFERENCES permissions(id))''')
+
+    cursor_per.execute('''INSERT INTO role_permissions (role_id, permission_id)
+                   VALUES ((SELECT id FROM roles WHERE role_name = 'admin'), (SELECT id FROM permissions WHERE permission_name = 'create')),
+                          ((SELECT id FROM roles WHERE role_name = 'admin'), (SELECT id FROM permissions WHERE permission_name = 'read')),
+                           ((SELECT id FROM roles WHERE role_name = 'admin'), (SElECT id FROM permissions WHERE permission_name = 'update')),
+                            ((SELECT id FROM roles WHERE role_name = 'admin'), (SELECT id FROM permissions WHERE permission_name = 'delete'))
+                    ''')
+    
+    cursor_per.execute('''INSERT INTO role_permissions (role_id, permission_id)
+                       VALUES ((SELECT id FROM roles WHERE role_name = 'user'), (SELECT id FROM permissions WHERE permission_name = 'read')),
+                                ((SELECT id FROM roles WHERE role_name = 'user'), (SELECT id FROM permissions WHERE permission_name = 'update'))
+                       ''')
+    
+    permissioncon.commit()
+    permissioncon.close()
+
+def check_permission(username, action):
+    user_con = userdb_connection()
+    perm_con = permissiondb_connection()
+    user_cursor = user_con.cursor()
+    perm_cursor = perm_con.cursor()
+
+    user_cursor.execute('''
+            SELECT r.role_name
+            FROM users u
+            JOIN roles r ON u.role_id = r.id
+            WHERE u.username = ?
+            ''', (username,))
+    
+    role = user_cursor.fetchone()
+
+    if role is None:
+        user_con.close()
+        raise PermissionDenied("Keine Berechtigung!")
+    main_menu()
+
+    return True
+
+class PermissionDenied(Exception):
     pass
+
+
+
+def current_user(username):
+    user_conn = userdb_connection()
+    cursor_user = user_conn.cursor()
+
+    cursor_user.execute("SELECT username FROM users WHERE username = ?", (username,))
+    result = current_user.fetchone()
+
+    if result:
+        return result[0]
+    else:
+        return None
+    
+    current_username = result[0]
+
+    user_conn.close()
+
 
 def login_menu():
     userconn = userdb_connection()
@@ -176,20 +347,28 @@ def login_menu():
 
 
 def main_menu():
+    current_username = current_user('username')
     while True:
-        choice = input("Bücher ausleihen (2)\nBücher zurückgeben (3)\nBücher hinzufügen (4)\nBücher löschen (5)\nBeenden (6)\nGib eine Zahl ein: ")
+        try:
+            choice = input("Bücher ausleihen (2)\nBücher zurückgeben (3)\nBücher hinzufügen (4)\nBücher löschen (5)\nBeenden (6)\nGib eine Zahl ein: ")
 
-        if choice == '1':
-            rent_book()
-        elif choice == '2':
-            return_book()
-        elif choice == '3':
-            add_book()
-        elif choice == '4':
-            delete_book()
-        elif choice == '5':
-            quit()
-        else:
+            if choice == '1':
+                check_permission(current_user, 'update')
+                rent_book()
+            elif choice == '2':
+                check_permission(current_user, 'update')
+                return_book()
+            elif choice == '3':
+                check_permission(current_user, 'create')
+                add_book()
+            elif choice == '4':
+                check_permission(current_user, 'delete')
+                delete_book()
+            elif choice == '5':
+                quit()
+            else:
+                print("Falsche Eingabe!")
+        except ValueError:
             print("Falsche Eingabe!")
 
 if __name__ == '__main__':
